Refactor opentitan into a bin+lib crate
The OpenTitan project is using the [bazel](https://bazel.build/) build
system and wants to depend on the tock repos as external resources.  We
are using [cargo raze](https://github.com/google/cargo-raze/) to process
`Cargo.toml` files from external projects and generate the appropriate
bazel BUILD files.  As such, we need to depend on the Tock `opentitan`
crate as a lib crate.
