+++ boards/opentitan/src/tests/rsa_4096.rs
@@ -2,9 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 // Copyright Tock Contributors 2022.
 
+use crate::setup::{PERIPHERALS, RSA_HARDWARE};
 use crate::tests::run_kernel_op;
-use crate::PERIPHERALS;
-use crate::RSA_HARDWARE;
 use capsules_extra::public_key_crypto::rsa_keys::RSA4096Keys;
 use core::cell::Cell;
 use kernel::hil::public_key_crypto::keys::{PubKey, PubPrivKey, RsaKey, RsaPrivKey};
