+++ boards/opentitan/src/tests/sip_hash.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 // Copyright Tock Contributors 2022.
 
+use crate::setup::SIPHASH;
 use crate::tests::run_kernel_op;
-use crate::SIPHASH;
 use core::cell::Cell;
 use kernel::hil::hasher::{self, Hasher};
 use kernel::static_init;
